Changed GPS meas type to doubles and put unit conversions in getMeas() methods for each sensor type

+++ include/Sensors/SensingSystem.h
@@ -15,7 +15,7 @@ class SensingSystem {
     //Make sure all of the sensors within sensors objects of one "group" use the same refresh rate 
     //(all of the sensors that contribute to one set of measurement must have the same refresh rate)
     Sensors<IMU_BNO055,double> IMUs_Avio;
-    Sensors<GPS_UBX9,long> GPSs_Avio;
+    Sensors<GPS_UBX9,double> GPSs_Avio;
     Sensors<BARO_BMP280,float> BAROs_Avio;
 
   public :
@@ -30,7 +30,7 @@ class SensingSystem {
     Sensors<IMU_BNO055,double>* getIMUs_Avio();
     uint8_t getIMUs_Avio_rl_amount();
 
-    Sensors<GPS_UBX9,long>* getGPSs_Avio(); 
+    Sensors<GPS_UBX9,double>* getGPSs_Avio(); 
     uint8_t getGPSs_Avio_rl_amount();
 
     Sensors<BARO_BMP280,float>* getBAROs_Avio();
@@ -40,7 +40,7 @@ class SensingSystem {
     double* getIMUs_CG_meas(); //Data combination from the entire sensing system to obtain IMU data at CG for use in the Kalman filter
     uint16_t getIMUs_CG_Hz(); //get the refresh rate of the IMUs that from the "CG" group
 
-    long* getGPSs_meas(); //Data combination from the entire sensing system to obtain GPS data for use in the Kalman filter
+    double* getGPSs_meas(); //Data combination from the entire sensing system to obtain GPS data for use in the Kalman filter
     uint16_t getGPSs_Hz(); //get the refresh rate of this group of GPSs
 
     float* getBAROs_meas(); //Data combination from the entire sensing system to obtain BARO data for use in Barometric System
